+++ src/main/java/com/crazycatzzz/client/DiscordWebSocket.java
@@ -1,0 +1,9 @@
+package com.crazycatzzz.client;
+
+/*
+ * WIP
+*/
+
+public class DiscordWebSocket {
+
+}
